+++ server/drivers/presto/index.js
@@ -34,9 +34,15 @@ function runQuery(query, connection) {
   let incomplete = false;
   const rows = [];
   const port = connection.port || 8080;
+  const protocol = connection.useHTTPS ? 'https' : 'http';
+  let url = `${connection.host}:${port}`;
+  if (!url.match(/^(http|https):\/\//i)) {
+    url = `${protocol}://${url}`;
+  }
   const prestoConfig = {
-    url: `http://${connection.host}:${port}`,
+    url,
     user: connection.username,
+    password: connection.password,
     catalog: connection.prestoCatalog,
     schema: connection.prestoSchema,
   };
@@ -101,6 +107,11 @@ const fields = [
     label: 'Database Username',
   },
   {
+    key: 'password',
+    formType: 'PASSWORD',
+    label: 'Database Password',
+  },
+  {
     key: 'prestoCatalog',
     formType: 'TEXT',
     label: 'Catalog',
@@ -109,6 +120,11 @@ const fields = [
     key: 'prestoSchema',
     formType: 'TEXT',
     label: 'Schema',
+  },
+  {
+    key: 'useHTTPS',
+    formType: 'CHECKBOX',
+    label: 'Use HTTPS',
   },
 ];