+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [6.6.0] - 2021-04-09
+
+- Add Presto authentication and HTTPS support
+- Fix ClickHouse queries beginning with comment
+- Update Dockerfile node and package dependencies
+
+## [6.5.0] - 2021-03-18
+
+- Add query run history side drawer
+- Add support for ClickHouse `LIMIT <offset_number>,<limit_number>` syntax
+
 ## [6.4.3] - 2021-03-14
 
 - Fix chart download button